+++ spec/spec_helper.rb
@@ -30,6 +30,10 @@ end
 
 require 'puppetlabs_spec_helper/module_spec_helper'
 
+if ENV['DEBUG']
+  Puppet::Util::Log.level = :debug
+  Puppet::Util::Log.newdestination(:console)
+end
 
 support_path = File.expand_path(File.join(File.dirname(__FILE__), '..','spec/support/*.rb'))
 Dir[support_path].each {|f| require f}